Implements data-validation feature requested in robin900/gspread-formatting#8.
Test coverage included.
